assignment/requirements.txt
# week1
python-dotenv
requests
lxml

# week3
numpy
matplotlib